# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AI-HRMS API
  version: 0.1.0
  description: >
    Contract index for the current AI-HRMS reproduction phase. The API serves
    one corporate group with multiple legal entities; legal entities are
    organization scopes, not SaaS tenants.
servers:
  - url: http://localhost:8080/api
    description: Local Go API
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Users]
      operationId: createUser
      summary: Create user
      description: Requires global role scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{id}:
    put:
      tags: [Users]
      operationId: updateUser
      summary: Update user
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/role-bindings:
    get:
      tags: [Roles]
      operationId: listUserRoleBindings
      summary: List user role bindings
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: User role bindings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBindingListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Roles]
      operationId: replaceUserRoleBindings
      summary: Replace user role bindings
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceRoleBindingsRequest"
      responses:
        "200":
          description: Updated user role bindings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBindingListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /employees:
    get:
      tags: [Employees]
      operationId: listEmployees
      summary: List employees with primary assignment
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged employees visible to caller scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeePageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Employees]
      operationId: createEmployee
      summary: Create employee
      description: Requires global role scope in the current implementation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /employees/export:
    get:
      tags: [Employees]
      operationId: exportEmployees
      summary: Export visible employees as CSV
      responses:
        "200":
          description: UTF-8 CSV file.
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /employees/{id}:
    get:
      tags: [Employees]
      operationId: getEmployee
      summary: Get employee by id
      description: Returns only employees visible to caller scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Employee detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Employees]
      operationId: updateEmployee
      summary: Update employee
      description: Requires global role scope in the current implementation.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "200":
          description: Updated employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /attendance:
    get:
      tags: [Attendance]
      operationId: listAttendance
      summary: List attendance records
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged attendance visible to caller scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendancePageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Attendance]
      operationId: createAttendance
      summary: Check in an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRequest"
      responses:
        "201":
          description: Created attendance record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /attendance/export:
    get:
      tags: [Attendance]
      operationId: exportAttendance
      summary: Export visible attendance records as CSV
      responses:
        "200":
          description: UTF-8 CSV file.
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /attendance/{id}/checkout:
    put:
      tags: [Attendance]
      operationId: checkoutAttendance
      summary: Check out an attendance record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated attendance record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /messages:
    get:
      tags: [Messages]
      operationId: listMessages
      summary: List messages
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Messages]
      operationId: createMessage
      summary: Create message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Created message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /messages/{id}/comments:
    get:
      tags: [Messages]
      operationId: listComments
      summary: List message comments
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Message comments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Messages]
      operationId: createComment
      summary: Create message comment
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Created comment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Caller lacks the required RBAC scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found or outside caller scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ApiEnvelope:
      type: object
      required: [success, code, message]
      properties:
        success:
          type: boolean
        code:
          type: integer
          example: 2001
        message:
          type: string
        data: {}
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [success, code, message]
          properties:
            success:
              const: false
            code:
              type: integer
              example: 4001
            message:
              type: string
...
    UserEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/User"
...
    RoleBindingListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/RoleBinding"
    EmployeeEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/Employee"
    AttendanceEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/Attendance"
    MessageEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/Message"
    CommentEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/Comment"
    UserPageEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/UserPage"
    EmployeePageEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/EmployeePage"
    AttendancePageEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/AttendancePage"
    MessagePageEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/MessagePage"
    CommentListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
...
    CreateUserRequest:
      type: object
      required: [mobile, username]
      properties:
        mobile:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
          default: password
    UpdateUserRequest:
      type: object
      required: [username, enableState]
      properties:
        username:
          type: string
        enableState:
          type: integer
          enum: [0, 1]
    ReplaceRoleBindingsRequest:
      type: object
      required: [bindings]
      properties:
        bindings:
          type: array
          items:
            $ref: "#/components/schemas/RoleBindingInput"
    RoleBindingInput:
      type: object
      required: [roleCode, scopeType, includeDescendants]
      properties:
        roleCode:
          type: string
        scopeType:
          type: string
          enum: [global, legal_entity, org_unit]
        scopeId:
          type: [string, "null"]
          format: uuid
        includeDescendants:
          type: boolean
          default: true
...
    EmployeeInput:
      type: object
      description: >
        Employee profile payload. Primary assignment is returned by reads;
        assignment write APIs will be split out when multi-assignment UI lands.
      properties:
        userId:
          type: [string, "null"]
          format: uuid
        employeeNo:
          type: string
        name:
          type: string
        mobile:
          type: string
        status:
          type: string
        sex:
          type: string
        dateOfBirth:
          type: string
        highestDegreeOfEducation:
          type: string
        nationalArea:
          type: string
        passportNo:
          type: string
        idNumber:
          type: string
        nativePlace:
          type: string
        nation:
          type: string
        englishName:
          type: string
        maritalStatus:
          type: string
        birthday:
          type: string
        zodiac:
          type: string
        age:
          type: string
        constellation:
          type: string
        bloodType:
          type: string
        domicile:
          type: string
        politicalOutlook:
          type: string
        qq:
          type: string
        wechat:
          type: string
        placeOfResidence:
          type: string
        postalAddress:
          type: string
        personalMailbox:
          type: string
        emergencyContact:
          type: string
        emergencyContactNumber:
          type: string
        bankCardNumber:
          type: string
        openingBank:
          type: string
        graduateSchool:
          type: string
        major:
          type: string
        homeCompany:
          type: string
        title:
          type: string
        resume:
          type: string
        isThereAnyCompetitionRestriction:
          type: string
        remarks:
          type: string
        primaryAssignment:
          anyOf:
            - $ref: "#/components/schemas/AssignmentInput"
            - type: "null"
    AssignmentInput:
      type: object
      properties:
        legalEntityId:
          type: [string, "null"]
          format: uuid
        orgUnitId:
          type: [string, "null"]
          format: uuid
        positionTitle:
          type: string
        startDate:
          type: string
          format: date
        employmentType:
          type: string
    CreateAttendanceRequest:
      type: object
      required: [employeeId]
      properties:
        employeeId:
          type: string
          format: uuid
        attendanceStatus:
          type: integer
          default: 1
        remarks:
          type: string
    CreateMessageRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        category:
          type: string
          default: general
        content:
          type: string
          description: HTML content must be sanitized before rendering.
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: Comment HTML must be sanitized before rendering.
    User:
      type: object
      required: [id, mobile, username, enableState, createdAt]
      properties:
        id:
          type: string
          format: uuid
        mobile:
          type: string
        username:
          type: string
        enableState:
          type: integer
        createdAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            type: string
...
    RoleBinding:
      allOf:
        - $ref: "#/components/schemas/RoleBindingInput"
        - type: object
          required: [id, userId, roleCode, scopeType, includeDescendants]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            roleId:
              type: string
              format: uuid
            roleName:
              type: string
            scopeName:
              type: string
            createdAt:
              type: string
              format: date-time
...
    Employee:
      allOf:
        - $ref: "#/components/schemas/EmployeeInput"
        - type: object
          required: [id, employeeNo, name, mobile, status]
          properties:
            id:
              type: string
              format: uuid
            primaryAssignment:
              anyOf:
                - $ref: "#/components/schemas/Assignment"
                - type: "null"
    Attendance:
      type: object
      required:
        - id
        - employeeId
        - employeeName
        - mobile
        - orgUnitName
        - attendanceStatus
        - day
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        employeeName:
          type: string
        mobile:
          type: string
        orgUnitName:
          type: string
        attendanceStatus:
          type: integer
        attendanceInTime:
          type: [string, "null"]
          format: date-time
        attendanceOutTime:
          type: [string, "null"]
          format: date-time
        attendanceInPlace:
          type: string
        day:
          type: string
        remarks:
          type: string
    Message:
      type: object
      required: [id, title, category, content, author, orgUnitName, star, view, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        category:
          type: string
        content:
          type: string
        author:
          type: string
        orgUnitName:
          type: string
        star:
          type: integer
        view:
          type: integer
        createdAt:
          type: string
          format: date-time
    Comment:
      type: object
      required: [id, messageId, content, username, createdAt]
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
        content:
          type: string
        username:
          type: string
        orgUnitName:
          type: string
        createdAt:
          type: string
          format: date-time
    UserPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            rows:
              type: array
              items:
                $ref: "#/components/schemas/User"
    EmployeePage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            rows:
              type: array
              items:
                $ref: "#/components/schemas/Employee"
    AttendancePage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            rows:
              type: array
              items:
                $ref: "#/components/schemas/Attendance"
    MessagePage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            rows:
              type: array
              items:
                $ref: "#/components/schemas/Message"
...